# perf(npm): optimize loading npm resolution snapshot from lockfile by only loading necessary version info