--- a/Cargo.toml
+++ b/Cargo.toml
@@ -54,7 +54,7 @@ deno_bench_util = { version = "0.175.0", path = "./bench_util" }
 deno_config = { version = "=0.39.3", features = ["workspace", "sync"] }
 deno_lockfile = "=0.23.2"
 deno_media_type = { version = "0.2.0", features = ["module_specifier"] }
-deno_npm = "=0.26.0"
+deno_npm = { version = "=0.26.0", path = "../deno_npm" }
 deno_path_util = "=0.2.1"
 deno_permissions = { version = "0.41.0", path = "./runtime/permissions" }
 deno_runtime = { version = "0.190.0", path = "./runtime" }