@dsherret 2025-04-15
I think it's safe to close this one now that we have lockfile v5?